Initial Commit
Based on https://github.com/Xaymar/teensy40-rs/ as well as a much older project attempting to get any use out of Teensyduino. Since the latter didn't actually make it anywhere, and none of the code is used, the Copyright date is accurate for the currently committed code.

The original project was made in Rust and I tested it with the classic Blinking LED project, something that Teensyduino required >200kb of Flash for. Rust only needed 86kb, however it was still way too large for my taste - and the endless hacking around limitations in Rust was annoying too.

This is yet another attempt at a clean lightweight bootloader for the Teensy 4.x series, but in C/C++. I'm not quite sure how small it'll be, but surely we can be within 10% of Rust's binary size.